+++ src/neptune/cli/containers.py
@@ -39,12 +39,12 @@ from typing import (
 )
 
 from neptune.cli.utils import get_qualified_name
-from neptune.common.exceptions import NeptuneConnectionLostException
 from neptune.constants import ASYNC_DIRECTORY
 from neptune.core.components.operation_storage import OperationStorage
 from neptune.core.components.queue.disk_queue import DiskQueue
 from neptune.envs import NEPTUNE_SYNC_BATCH_TIMEOUT_ENV
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import NeptuneConnectionLostException
 from neptune.internal.id_formats import UniqueId
 from neptune.internal.operation import Operation
 from neptune.internal.operation_processors.utils import get_container_dir
