+++ src/neptune/cli/utils.py
@@ -36,7 +36,6 @@ from typing import (
     Union,
 )
 
-from neptune.common.exceptions import NeptuneException
 from neptune.core.components.queue.disk_queue import DiskQueue
 from neptune.envs import PROJECT_ENV_NAME
 from neptune.exceptions import (
@@ -49,6 +48,7 @@ from neptune.internal.backends.api_model import (
 )
 from neptune.internal.backends.neptune_backend import NeptuneBackend
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import NeptuneException
 from neptune.internal.id_formats import (
     QualifiedName,
     UniqueId,
