+++ src/neptune/common/oauth.py
@@ -27,8 +27,8 @@ from requests.auth import AuthBase
 from requests_oauthlib import OAuth2Session
 
 from neptune.common.backends.utils import with_api_exceptions_handler
-from neptune.common.exceptions import NeptuneInvalidApiTokenException
 from neptune.common.utils import update_session_proxies
+from neptune.internal.exceptions import NeptuneInvalidApiTokenException
 
 _decoding_options = {
     "verify_signature": False,
