+++ src/neptune/common/utils.py
@@ -26,7 +26,9 @@ import numpy as np
 import pandas as pd
 
 from neptune.common import envs
-from neptune.common.exceptions import (
+from neptune.common.git_info import GitInfo
+from neptune.common.patterns import PROJECT_QUALIFIED_NAME_PATTERN
+from neptune.internal.exceptions import (
     FileNotFound,
     InvalidNotebookPath,
     NeptuneIncorrectProjectQualifiedNameException,
@@ -34,8 +36,6 @@ from neptune.common.exceptions import (
     NotADirectory,
     NotAFile,
 )
-from neptune.common.git_info import GitInfo
-from neptune.common.patterns import PROJECT_QUALIFIED_NAME_PATTERN
 from neptune.internal.utils.logger import get_logger
 
 _logger = get_logger()
