+++ src/neptune/internal/backends/neptune_backend.py
@@ -26,7 +26,6 @@ from typing import (
 )
 
 from neptune.api.dtos import FileEntry
-from neptune.common.exceptions import NeptuneException
 from neptune.core.components.operation_storage import OperationStorage
 from neptune.internal.artifacts.types import ArtifactFileData
 from neptune.internal.backends.api_model import (
@@ -52,6 +51,7 @@ from neptune.internal.backends.api_model import (
 )
 from neptune.internal.backends.nql import NQLQuery
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import NeptuneException
 from neptune.internal.id_formats import (
     QualifiedName,
     UniqueId,
