+++ src/neptune/internal/backends/neptune_backend_mock.py
@@ -35,10 +35,6 @@ from typing import (
 from zipfile import ZipFile
 
 from neptune.api.dtos import FileEntry
-from neptune.common.exceptions import (
-    InternalClientError,
-    NeptuneException,
-)
 from neptune.core.components.operation_storage import OperationStorage
 from neptune.exceptions import (
     ContainerUUIDNotFound,
@@ -76,6 +72,10 @@ from neptune.internal.backends.neptune_backend import NeptuneBackend
 from neptune.internal.backends.nql import NQLQuery
 from neptune.internal.container_structure import ContainerStructure
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import (
+    InternalClientError,
+    NeptuneException,
+)
 from neptune.internal.id_formats import (
     QualifiedName,
     SysId,
