+++ src/neptune/internal/backends/operation_api_name_visitor.py
@@ -15,7 +15,7 @@
 #
 __all__ = ["OperationApiNameVisitor"]
 
-from neptune.common.exceptions import InternalClientError
+from neptune.internal.exceptions import InternalClientError
 from neptune.internal.operation import (
     AddStrings,
     AssignArtifact,
