+++ src/neptune/internal/backends/operation_api_object_converter.py
@@ -15,7 +15,7 @@
 #
 __all__ = ["OperationApiObjectConverter"]
 
-from neptune.common.exceptions import InternalClientError
+from neptune.internal.exceptions import InternalClientError
 from neptune.internal.operation import (
     AddStrings,
     AssignArtifact,
