+++ src/neptune/internal/operation.py
@@ -28,13 +28,13 @@ from typing import (
     TypeVar,
 )
 
-from neptune.common.exceptions import (
-    InternalClientError,
-    NeptuneException,
-)
 from neptune.core.components.operation_storage import OperationStorage
 from neptune.exceptions import MalformedOperation
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import (
+    InternalClientError,
+    NeptuneException,
+)
 from neptune.internal.types.file_types import FileType
 from neptune.types.atoms.file import File
