+++ src/neptune/internal/operation_processors/async_operation_processor.py
@@ -33,7 +33,6 @@ from typing import (
     Tuple,
 )
 
-from neptune.common.exceptions import NeptuneException
 from neptune.common.warnings import (
     NeptuneWarning,
     warn_once,
@@ -45,6 +44,7 @@ from neptune.core.components.operation_storage import OperationStorage
 from neptune.core.components.queue.disk_queue import DiskQueue
 from neptune.envs import NEPTUNE_SYNC_AFTER_STOP_TIMEOUT
 from neptune.exceptions import NeptuneSynchronizationAlreadyStoppedException
+from neptune.internal.exceptions import NeptuneException
 from neptune.internal.init.parameters import DEFAULT_STOP_TIMEOUT
 from neptune.internal.operation import Operation
 from neptune.internal.operation_processors.operation_logger import ProcessorStopLogger
