+++ src/neptune/metadata_containers/model.py
@@ -26,7 +26,6 @@ from typing import (
 from typing_extensions import Literal
 
 from neptune.attributes.constants import SYSTEM_NAME_ATTRIBUTE_PATH
-from neptune.common.exceptions import NeptuneException
 from neptune.envs import CONNECTION_MODE
 from neptune.exceptions import (
     InactiveModelException,
@@ -44,6 +43,7 @@ from neptune.internal.backends.nql import (
     NQLQueryAttribute,
 )
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import NeptuneException
 from neptune.internal.id_formats import QualifiedName
 from neptune.internal.init.parameters import (
     ASYNC_LAG_THRESHOLD,
