+++ src/neptune/metadata_containers/model_version.py
@@ -28,7 +28,6 @@ from neptune.attributes.constants import (
     SYSTEM_NAME_ATTRIBUTE_PATH,
     SYSTEM_STAGE_ATTRIBUTE_PATH,
 )
-from neptune.common.exceptions import NeptuneException
 from neptune.envs import CONNECTION_MODE
 from neptune.exceptions import (
     InactiveModelVersionException,
@@ -38,6 +37,7 @@ from neptune.exceptions import (
 )
 from neptune.internal.backends.api_model import ApiExperiment
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import NeptuneException
 from neptune.internal.id_formats import QualifiedName
 from neptune.internal.init.parameters import (
     ASYNC_LAG_THRESHOLD,
