+++ tests/unit/neptune/new/client/test_model_version.py
@@ -23,7 +23,6 @@ from neptune import (
     init_model_version,
 )
 from neptune.attributes import String
-from neptune.common.exceptions import NeptuneException
 from neptune.common.warnings import (
     NeptuneWarning,
     warned_once,
@@ -44,6 +43,7 @@ from neptune.internal.backends.api_model import (
 )
 from neptune.internal.backends.neptune_backend_mock import NeptuneBackendMock
 from neptune.internal.container_type import ContainerType
+from neptune.internal.exceptions import NeptuneException
 from tests.unit.neptune.new.client.abstract_experiment_test_mixin import AbstractExperimentTestMixin
 from tests.unit.neptune.new.utils.api_experiments_factory import (
     api_model,
